+++ featuretools/__init__.py
@@ -1,5 +1,4 @@
 # flake8: noqa
-import warnings
 from .config_init import config
 from . import variable_types
 from .entityset.api import *
@@ -15,8 +14,13 @@ import featuretools.demo
 from . import feature_base
 from .feature_base import AggregationFeature, DirectFeature, Feature, FeatureBase, IdentityFeature, TransformFeature, GroupByTransformFeature, save_features, load_features
 
+import logging
 import pkg_resources
 import sys
+import traceback
+
+logger = logging.getLogger('featuretools')
+
 # Call functions registered by other libraries when featuretools is imported
 for entry_point in pkg_resources.iter_entry_points('featuretools_initialize'):
     try:
@@ -31,4 +35,7 @@ for entry_point in pkg_resources.iter_entry_points('featuretools_plugin'):
     try:
         sys.modules["featuretools." + entry_point.name] = entry_point.load()
     except Exception:
-        pass
+        message = "Featuretools failed to load plugin {} from library {}. "
+        message += "For a full stack trace, set logging to debug."
+        logger.warn(message.format(entry_point.name, entry_point.module_name))
+        logger.debug(traceback.format_exc())