+++ docs/source/changelog.rst
@@ -3,6 +3,8 @@
 Changelog
 ---------
 **Future Release**
+    * Changes
+        * Added Plugin Warnings at Entry Point (:pr:`850`)
     * Documentation Changes
         * Add links to primitives.featurelabs.com (:pr:`860`)
         * Add source code links to API reference (:pr:`862`)
@@ -10,7 +12,7 @@ Changelog
         * Miscellaneous changes (:pr:`861`)
 
     Thanks to the following people for contributing to this release:
-    :user:`rwedge`, :user:`thehomebrewnerd`
+    :user:`rwedge`, :user:`thehomebrewnerd`, :user:`jeff-hernandez`
 
 **v0.13.3 Feb 28, 2020**
     * Fixes